connect wallet works!
