+++ src/context/Web3Context.tsx
@@ -4,8 +4,34 @@ import React, { type ReactNode } from 'react'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { cookieToInitialState, WagmiProvider, type Config } from 'wagmi'
 
-import { wagmiAdapter } from '@/lib/web3/wagmi'
-import { modal, metadata } from '@/lib/web3/appkit'
+import { createAppKit } from '@reown/appkit'
+
+import { wagmiAdapter, projectId, networks } from '@/lib/web3/wagmi'
+import { rsktestnet } from '@/lib/web3/chains/RootstockTestnet';
+
+if (!projectId) {
+  throw new Error('Project ID is not defined')
+}
+
+// Set up metadata
+const metadata = {
+  name: 'Dash Mon[k]ey',
+  description: 'Monkey see, Monkey do!',
+  url: 'http://localhost:3000', // origin must match your domain & subdomain
+  icons: ['/circus-monkey.svg']
+}
+
+// Create the modal
+export const modal = createAppKit({
+  adapters: [wagmiAdapter],
+  projectId,
+  networks,
+  defaultNetwork: rsktestnet,
+  metadata: metadata,
+  features: {
+    analytics: true // Optional - defaults to your Cloud configuration
+  }
+});
 
 // Set up queryClient
 const queryClient = new QueryClient()
