+++ src/lib/web3/wagmi.ts
@@ -1,5 +1,6 @@
 import { cookieStorage, createStorage } from '@wagmi/core'
 import { WagmiAdapter } from '@reown/appkit-adapter-wagmi'
+import type { AppKitNetwork } from '@reown/appkit/networks';
 
 import { rsktestnet } from './chains/RootstockTestnet';
 
@@ -10,7 +11,7 @@ if (!projectId) {
   throw new Error('Project ID is not defined');
 }
 
-export const networks = [rsktestnet]
+export const networks = [rsktestnet] as [AppKitNetwork, ...AppKitNetwork[]];
 
 //Set up the Wagmi Adapter (Config)
 export const wagmiAdapter = new WagmiAdapter({
